fix gpu assignment in local-cluster mode and bash cli's (#973)
Fixes gpu assignment in local-cluster mode (used in test of no-code
change cli's) to match what is done local mode.

This gpu assignment error was silently failing in the cli runs in
run_test.sh in CI runs because the cli's weren't propagating Spark
errors to their exit codes. This PR also fixes this issue in the clis by
setting `check=True` in the Python `subprocess.run` invocations and adds
a test to check the Spark errors are propagated correctly.

Signed-off-by: Erik Ordentlich <eordentlich@gmail.com>