correctly handle corner case barrier rdd errors  (#993)
We currently have an optimization that if input dfs have the same number
of partitions as num_workers, we skip the repartition. This can be
problematic in cases when the df has ancestors with a different number
of partitions before a shuffle boundary, which is incompatible with
barrier rdd logic. There are two fixes:
- In the case of the algos that collect the barrier rdd results to get
model training results, the default path is taken and is retried with a
hash repartitioning (introducing a shuffle boundary) if a barrier error
is encountered.
- In the case of knn and dbscan which return the barrier rdd after
converting to a df, there is no opportunity to retry since the calling
user code would encounter this error. In this case, the only option is
to repartition the input in all cases.

---------

Signed-off-by: Erik Ordentlich <eordentlich@gmail.com>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>